rename resolve pkg to resolver
