+++ pkg/notify/resolver/resolver.go
@@ -1,4 +1,4 @@
-package resolve
+package resolver
 
 import (
 	"bufio"
